Drop `ucx_proc_type` variant and use Jinja instead
When adding the CUDA versions, it is already possible and useful to
handle cases where a CPU-only build is generated. Having this variant in
addition to CUDA versions results in an overcomplete spanning of this
space. So eliminate `ucx_proc_type` as a variant and instead set it in
Jinja as a convenience for use in the recipe.