Fill in the `cuda` compiler details
Make sure that `nvcc` is required and that all supported versions are
used for the builds. This is added to the feedstock temporarily.
However the goal is to integrate this into conda-forge-pinning over the
long term. So that all feedstocks will benefit from the configuration
automatically and individual feedstocks merely need to use the `cuda`
compiler as part of their build process.