Hackily add `zip_key` variant to `ucx_proc`
Conda-build tries to validate the `zip_keys` are use by all packages.
However there is no need for `ucx-proc` to use the `zip_keys` we've
added. To fix that issue, we make a small hack where we include an
`extra` section for the `ucx-proc` package where the
`cuda_compiler_version` is included. This has no practical impact
outside of making conda-build happy. Would be good to get conda-build to
lighten up here, but this will do for now.