Try not specifying the path to rdmacm
This was originally needed to build the package when the recipe was
first being updated. Though it may not be strictly necessary at this
stage. In particular CDTs have been built out and added as dependencies
for rdmacm (amongst other system level dependencies that UCX has). So
try dropping this explicit path to see if things still work.