Drop unneeded commented code
There is no need to run `autogen.sh` on published release artifacts
where this has already been done for us. So this line has remained
commented. It's only really needed for dev builds where we already add
this line. So in the interest of keeping things easy to understand, drop
this already commented line.