Drop the custom build string from UCX packages
As the CUDA version and whether the package is CPU or GPU is already
included in the package hash correctly, go ahead and drop the custom
build strings in packages. This info can be determined by looking at the
`cudatoolkit` version and `ucx-proc` package installed. Alternatively
they can be determined by running `conda inspect hash-inputs` on the
packages. Dropping the custom build strings should make this a bit
easier to maintain.