Build ucx-py only under CUDA 10.2
As we are currently building `ucx-py` for every CUDA version (including
when CUDA is disabled), we produce the same packages multiple times. If
we have a race condition on uploading, Anaconda.org doesn't block the
package from being uploaded more than once. This causes issues when
installing the package. To fix these issues, make sure we only build
ucx-py in the CUDA 10.2 builds. The choice is really arbitrary as it
does not matter. In any event this should ensure each ucx-py build only
occurs once (per Python version) so only one package is uploaded to
Anaconda.org, which should avoid this issue coming up again.